+++ addOns/reports/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+- Support for template sections
+- Automation job: support risk, confidence and section configuration
 
 ## [0.1.0] - 2021-03-19